# client/openapi.json
{"openapi":"3.1.0","info":{"title":"DAG Todo API","description":"Task management with DAG dependencies and real-time updates","version":"0.1.0"},"paths":{"/api/tasks/subscribe":{"get":{"summary":"Subscribe Tasks","description":"Subscribe to real-time task updates via SSE.","operationId":"subscribe_tasks_api_tasks_subscribe_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/tasks":{"get":{"summary":"List Tasks","description":"List all tasks with computed properties.","operationId":"list_tasks_api_tasks_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NodeListOut"}}}}}},"post":{"summary":"Add Task","description":"Create a new task.","operationId":"add_task_api_tasks_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NodeCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NodeOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/tasks/{task_id}":{"get":{"summary":"Get Task","description":"Get a single task with computed properties.","operationId":"get_task_api_tasks__task_id__get","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NodeOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"summary":"Set Task","description":"Update a task's properties.","operationId":"set_task_api_tasks__task_id__patch","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NodeUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OperationResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Remove Task","description":"Delete a task and its edges.","operationId":"remove_task_api_tasks__task_id__delete","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OperationResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/tasks/{task_id}/rename":{"post":{"summary":"Rename Task","description":"Rename a task (change its ID).","operationId":"rename_task_api_tasks__task_id__rename_post","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RenameRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OperationResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/links":{"post":{"summary":"Link Tasks","description":"Create a dependency: from_id depends on to_id.","operationId":"link_tasks_api_links_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LinkRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DependencyOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Unlink Tasks","description":"Remove a dependency.","operationId":"unlink_tasks_api_links_delete","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LinkRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OperationResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/init":{"post":{"summary":"Init Db","description":"Initialize the database schema and run migrations.","operationId":"init_db_api_init_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OperationResult"}}}}}}},"/health":{"get":{"summary":"Health","description":"Health check endpoint.","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"DependencyOut":{"properties":{"id":{"type":"string","title":"Id"},"from_id":{"type":"string","title":"From Id"},"to_id":{"type":"string","title":"To Id"}},"type":"object","required":["id","from_id","to_id"],"title":"DependencyOut","description":"Dependency relationship."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LinkRequest":{"properties":{"from_id":{"type":"string","title":"From Id"},"to_id":{"type":"string","title":"To Id"}},"type":"object","required":["from_id","to_id"],"title":"LinkRequest","description":"Link/unlink request."},"NodeCreate":{"properties":{"text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Text"},"completed":{"type":"boolean","title":"Completed","default":false},"node_type":{"$ref":"#/components/schemas/NodeType","default":"Task"},"due":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Due"},"id":{"type":"string","title":"Id"},"depends":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Depends"},"blocks":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Blocks"}},"type":"object","required":["id"],"title":"NodeCreate","description":"Node creation request."},"NodeListOut":{"properties":{"tasks":{"additionalProperties":{"$ref":"#/components/schemas/NodeOut"},"type":"object","title":"Tasks"},"dependencies":{"additionalProperties":{"$ref":"#/components/schemas/DependencyOut"},"type":"object","title":"Dependencies"},"has_cycles":{"type":"boolean","title":"Has Cycles"}},"type":"object","required":["tasks","dependencies","has_cycles"],"title":"NodeListOut","description":"Node list response."},"NodeOut":{"properties":{"id":{"type":"string","title":"Id"},"text":{"type":"string","title":"Text"},"node_type":{"$ref":"#/components/schemas/NodeType"},"completed":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Completed"},"due":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Due"},"created_at":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Created At"},"updated_at":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Updated At"},"calculated_value":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Calculated Value"},"calculated_due":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Calculated Due"},"deps_clear":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Deps Clear"},"is_actionable":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Actionable"},"parents":{"items":{"type":"string"},"type":"array","title":"Parents"},"children":{"items":{"type":"string"},"type":"array","title":"Children"}},"type":"object","required":["id","text","node_type","completed","due","created_at","updated_at","calculated_value","calculated_due","deps_clear","is_actionable","parents","children"],"title":"NodeOut","description":"Node response with calculated fields."},"NodeType":{"type":"string","enum":["Task","And","Or","Not","ExactlyOne"],"title":"NodeType","description":"Node type labels."},"NodeUpdate":{"properties":{"text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Text"},"completed":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Completed"},"node_type":{"anyOf":[{"$ref":"#/components/schemas/NodeType"},{"type":"null"}]},"due":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Due"}},"type":"object","title":"NodeUpdate","description":"Node update request."},"OperationResult":{"properties":{"success":{"type":"boolean","title":"Success"},"message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Message"}},"type":"object","required":["success"],"title":"OperationResult","description":"Generic operation result."},"RenameRequest":{"properties":{"new_id":{"type":"string","title":"New Id"}},"type":"object","required":["new_id"],"title":"RenameRequest","description":"Rename request."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}